refactor(vectordb): extract vector db factory to reduce code duplication
Move PatchedQdrant and AgnoDoc implementations to a shared factory module
Update agent implementations to use the new factory pattern
