+++ chat_cli/app/vectordb/__init__.py
@@ -1,0 +1,1 @@
+# vectordb package
